+++ CHANGELOG.md
@@ -11,11 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 -   Project Name can be specified for merge filter #394
 
 ### Changed
+-   Throw a MergeException if project names do not match in MergeFilter #394 
 
 ### Removed
 
 ### Fixed
--   Throw a MergeException if project names do not match in MergeFilter #394 
 
 ### Chore